Revert "Employtics re vamp"
